third_party/blink/web_tests/fast/css/insertRule-media-expected.txt

Test parsing of @media rule using insertRule(). http://bugs.webkit.org/show_bug.cgi?id=15986

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.styleSheets[0].insertRule(rule, 0) is 0
PASS document.styleSheets[0].rules[0].cssText is serializedRule
PASS successfullyParsed is true

TEST COMPLETE
